fix: restructure for playground components
