+++ apps/extension/src/containers/playground/containers/accounts/account-activity/account-activity.css.ts
