+++ apps/extension/src/containers/playground/containers/accounts/account-switcher/account-switcher.css.ts
