+++ apps/extension/src/containers/playground/containers/accounts/accounts-home/accounts-home.css.ts
