+++ apps/extension/src/containers/playground/containers/accounts/accounts-home/index.tsx
@@ -5,10 +5,10 @@ import { PlusIcon, MagnifyingGlassIcon } from 'ui/src/components/icons'
 // import clsx from 'clsx'
 // import { BrowserRouter, Routes, Route, Link, useLocation, useMatch } from 'react-router-dom'
 import { Link, useMatch } from 'react-router-dom'
-import { DropdownProfile } from '@src/containers/playground/containers/accounts/components/dropdown-profile'
-import { AccountsList } from '@src/containers/playground/containers/accounts/components/accounts-list'
-import { AccountSwitcher } from '@src/containers/playground/containers/accounts/components/account-switcher'
-import { AccountActivity } from '@src/containers/playground/containers/accounts/components/account-activity'
+import { DropdownProfile } from '@src/containers/playground/containers/accounts/dropdown-profile'
+import { AccountsList } from '@src/containers/playground/containers/accounts/accounts-list'
+import { AccountSwitcher } from '@src/containers/playground/containers/accounts/account-switcher'
+import { AccountActivity } from '@src/containers/playground/containers/accounts/account-activity'
 import { useAccountParams } from '@src/containers/playground/hooks/use-account-params'
 import { motion } from 'framer-motion'
 import { Button } from 'ui/src/components-v2/button'
