+++ apps/extension/src/containers/playground/containers/accounts/accounts-list/accounts-list.css.ts
