+++ apps/extension/src/containers/playground/containers/accounts/desktop/accounts-desktop.css.ts
