+++ apps/extension/src/containers/playground/containers/accounts/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
-import { AccountsDesktop } from './components/desktop'
-import { AccountsMobile } from './components/mobile'
+import { AccountsDesktop } from './desktop'
+import { AccountsMobile } from './mobile'
 
 export const Accounts: React.FC = () => (
 	<>
