+++ apps/extension/src/containers/playground/containers/accounts/navigation/navigation.css.ts
